3rd Revision
Cleaned up order form code, re-factored database: The columns are now in
the order they are on the paper with the exception of attachments being
the last column, added click events to the two buttons in the accountant
page.  I will add methods to those buttons in the next revision.