Changed header to Login/out, redirects, css
Added link to request form and added functionality that changes Login
button to message displaying name and a logout button when logged in.

Put in redirects from login to defined pages, and from secure pages
(Purchases only for now) to login if not logged in, or to the
appropriate page if logged in, but accessing other pages.

Added css tags to login, purchases page.