Place/Cancel button, Fixed edit, default
Added a place order button to Purchasers, which records the time and the
name of the purchaser placing it. This button turns into a cancel order
button if an order exists.

Added missing UserID field to default.aspx.cs. It defaults to null.

Fixed autofilling of boolean statements in Editorder.aspx.cs.